+++ src/main/java/com/LetMeDoWith/LetMeDoWith/domain/task/service/strategy/TodoTaskRoutineDateCalculateStrategy.java
@@ -1,4 +1,4 @@
-package com.LetMeDoWith.LetMeDoWith.domain.task.service;
+package com.LetMeDoWith.LetMeDoWith.domain.task.service.strategy;
 
 import com.LetMeDoWith.LetMeDoWith.domain.task.enums.TodoTaskRoutineCycle;
 import java.time.LocalDate;
