+++ src/test/java/com/LetMeDoWith/LetMeDoWith/application/task/service/RegisterTodoTaskServiceTest.java
@@ -18,9 +18,9 @@ import com.LetMeDoWith.LetMeDoWith.common.exception.status.FailResponseStatus;
 import com.LetMeDoWith.LetMeDoWith.domain.task.enums.TodoTaskRoutineCycle;
 import com.LetMeDoWith.LetMeDoWith.domain.task.model.TaskCategory;
 import com.LetMeDoWith.LetMeDoWith.domain.task.model.TodoTask;
-import com.LetMeDoWith.LetMeDoWith.domain.task.service.DailyRoutineDateCalculateStrategy;
-import com.LetMeDoWith.LetMeDoWith.domain.task.service.TodoTaskRoutineDateCalculateStrategy;
 import com.LetMeDoWith.LetMeDoWith.domain.task.service.TodoTaskRoutineDateCalculator;
+import com.LetMeDoWith.LetMeDoWith.domain.task.service.strategy.DailyRoutineDateCalculateStrategy;
+import com.LetMeDoWith.LetMeDoWith.domain.task.service.strategy.TodoTaskRoutineDateCalculateStrategy;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
