+++ src/test/java/com/LetMeDoWith/LetMeDoWith/domain/task/service/TodoTaskRoutineDateCalculatorTest.java
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.when;
 import com.LetMeDoWith.LetMeDoWith.common.exception.RestApiException;
 import com.LetMeDoWith.LetMeDoWith.common.exception.status.FailResponseStatus;
 import com.LetMeDoWith.LetMeDoWith.domain.task.enums.TodoTaskRoutineCycle;
+import com.LetMeDoWith.LetMeDoWith.domain.task.service.strategy.TodoTaskRoutineDateCalculateStrategy;
 import java.time.LocalDate;
 import java.util.Map;
 import java.util.Set;
